backend/requirements-dev.txt
-r requirements.txt
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-docker>=3.1.0
anyio[asyncio]>=4.0.0
testcontainers[postgres]>=4.8.0
httpx>=0.27.0
